modules/_scraper_template/requirements.txt
beautifulsoup4
requests
python-dotenv
pillow
supabase